requirements.txt
amqp==2.6.1
bidict==0.21.2
billiard==3.6.4.0
celery==4.4.7
cffi==1.14.5
click==7.1.2
Flask==1.1.2
Flask-SocketIO==5.0.1
flower==0.9.7
gevent==21.1.2
gevent-websocket==0.10.1
greenlet==1.0.0
humanize==3.4.1
importlib-metadata==3.3.0
itsdangerous==1.1.0
Jinja2==2.11.2
kombu==4.6.11
MarkupSafe==1.1.1
pip==20.1.1
prometheus-client==0.8.0
pycparser==2.20
python-engineio==4.1.0
python-socketio==5.2.1
pytz==2020.4
redis==3.5.3
setuptools==47.1.0
tornado==6.1
typing-extensions==3.7.4.3
vine==1.3.0
Werkzeug==1.0.1
zipp==3.4.0
zope.event==4.5.0
zope.interface==5.2.0
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation
opentelemetry-instrumentation-celery
opentelemetry-instrumentation-flask
opentelemetry-instrumentation-requests
opentelemetry-instrumentation-wsgi
opentelemetry-exporter-otlp-proto-grpc